+++ controllers/resource/requirementsUpdater.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/cloudogu/k8s-dogu-operator/controllers/localregistry"
+
 	v1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
@@ -13,6 +13,7 @@ import (
 	"github.com/cloudogu/cesapp-lib/registry"
 	k8sv1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
 	"github.com/cloudogu/k8s-dogu-operator/internal/cloudogu"
+	"github.com/cloudogu/k8s-registry-lib/dogu/local"
 	coreosclient "go.etcd.io/etcd/client/v2"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -27,7 +28,7 @@ type requirementsUpdater struct {
 	client            client.Client
 	namespace         string
 	registry          registry.Registry
-	localDoguRegistry localregistry.LocalDoguRegistry
+	localDoguRegistry local.LocalDoguRegistry
 	requirementsGen   cloudogu.ResourceRequirementsGenerator
 }
 
@@ -48,7 +49,7 @@ func NewRequirementsUpdater(client client.Client, namespace string, clientSet ku
 		client:    client,
 		namespace: namespace,
 		registry:  reg,
-		localDoguRegistry: localregistry.NewCombinedLocalDoguRegistry(
+		localDoguRegistry: local.NewCombinedLocalDoguRegistry(
 			clientSet.CoreV1().ConfigMaps(namespace),
 			reg,
 		),