+++ internal/cloudogu/cesregistry.go
@@ -30,22 +30,3 @@ type DoguRegistrator interface {
 	// UnregisterDogu removes a registration of a dogu from the local dogu registry.
 	UnregisterDogu(ctx context.Context, dogu string) error
 }
-
-// LocalDoguRegistry abstracts accessing various backends for reading and writing dogu specs (dogu.json).
-type LocalDoguRegistry interface {
-	// Enable makes the dogu spec reachable.
-	Enable(ctx context.Context, dogu *cesappcore.Dogu) error
-	// Register adds the given dogu spec to the local registry.
-	Register(ctx context.Context, dogu *cesappcore.Dogu) error
-	// UnregisterAllVersions deletes all versions of the dogu spec from the local registry and makes the spec unreachable.
-	UnregisterAllVersions(ctx context.Context, simpleDoguName string) error
-	// Reregister adds the new dogu spec to the local registry, enables it, and deletes all specs referenced by the old dogu name.
-	// This is used for namespace changes and may contain an empty implementation if this action is not necessary.
-	Reregister(ctx context.Context, newDogu *cesappcore.Dogu) error
-	// GetCurrent retrieves the spec of the referenced dogu's currently installed version.
-	GetCurrent(ctx context.Context, simpleDoguName string) (*cesappcore.Dogu, error)
-	// GetCurrentOfAll retrieves the specs of all dogus' currently installed versions.
-	GetCurrentOfAll(ctx context.Context) ([]*cesappcore.Dogu, error)
-	// IsEnabled checks if the current spec of the referenced dogu is reachable.
-	IsEnabled(ctx context.Context, simpleDoguName string) (bool, error)
-}