+++ internal/thirdParty/external.go
@@ -4,6 +4,7 @@ package thirdParty
 
 import (
 	"github.com/go-logr/logr"
+
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes"
 	appsv1client "k8s.io/client-go/kubernetes/typed/apps/v1"
@@ -15,6 +16,7 @@ import (
 
 	"github.com/cloudogu/cesapp-lib/registry"
 	"github.com/cloudogu/cesapp-lib/remote"
+	"github.com/cloudogu/k8s-registry-lib/dogu/local"
 )
 
 type K8sClient interface {
@@ -61,6 +63,11 @@ type ConfigurationContext interface {
 
 type ConfigurationRegistry interface {
 	registry.Registry
+}
+
+// LocalDoguRegistry abstracts accessing various backends for reading and writing dogu specs (dogu.json).
+type LocalDoguRegistry interface {
+	local.LocalDoguRegistry
 }
 
 type DeploymentInterface interface {